pyproject.toml
[project]
name = "pharmacokinetics"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "diffrax>=0.7.2",
    "equinox>=0.13.4",
    "evosax>=0.2.0",
    "gsax",
    "jax>=0.6.2",
    "marimo[recommended]>=0.20.4",
    "matplotlib>=3.10.3",
    "nbconvert>=7.17.0",
    "numpy>=2.3.1",
    "openpyxl>=3.1.5",
    "optax>=0.2.5",
    "pandas>=2.3.0",
    "playwright>=1.58.0",
    "pyarrow>=18.0.0",
    "pyfonts>=1.1.1",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "scikit-learn>=1.7.0",
    "scikit-optimize>=0.10.2",
    "seaborn>=0.13.2",
    "xarray>=2026.2.0",
]


[tool.ruff]
line-length = 128
select = ["E4", "E7", "E9", "F", "B"]
ignore = ["E501", "W292", "F401"]

[tool.uv.sources]
diffrax = { git = "https://github.com/patrick-kidger/diffrax" }
gsax = { git = "https://github.com/danielepessina/gsax", rev = "5bd6ed1ebbcd084420bcd7d437ab6e3955ceeb1c" }

[tool.uv.sources.torch]
index = "pytorch-cpu"

[tool.uv.sources.torchvision]
index = "pytorch-cpu"

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
explicit = true

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[tool.setuptools]
packages = ["pharmacokinetics"]
